# docs/openapi/service-marketplace-v1.yaml
openapi: 3.0.3
info:
  title: ReportPortal Plugin Marketplace Registry
  description: |
    REST API for the ReportPortal Plugin Marketplace registry (`service-marketplace`).

    Public read endpoints are unauthenticated and CDN-cacheable (except artifact download,
    which hits the registry origin). Operator endpoints require a bearer token: either a
    registry session JWT (Operator UI, ADR-010) or a GitHub Actions OIDC token (CI/CD, ADR-014).

    API path versioning follows ADR-006. This document describes `/api/v1/`.
  contact:
    name: Support
    email: support@reportportal.io
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  version: 1.0.0
servers:
  - url: https://marketplace.reportportal.io
    description: ReportPortal-operated registry (default)
  - url: '{protocol}://{host}'
    description: Self-hosted registry instance (ADR-004)
    variables:
      protocol:
        default: https
        enum:
          - http
          - https
      host:
        default: localhost:8080
        description: Registry host name and optional port
...
paths:
  /api/v1/plugins:
    get:
      tags:
        - Public — Plugins
      summary: List plugins
      description: |
        Returns all plugins from the precomputed registry index. Supports filtering by
        category and free-text search (FR-U-01). Removed plugins are excluded from the listing.

        Pagination is not supported in v1; the full catalogue is returned. Future additive
        pagination will not require a new API version (ADR-006).
      operationId: listPlugins
      parameters:
        - $ref: '#/components/parameters/CategoryQuery'
        - $ref: '#/components/parameters/SearchQuery'
      responses:
        '200':
          description: Plugin catalogue
          headers:
            Deprecation:
              $ref: '#/components/headers/Deprecation'
            Sunset:
              $ref: '#/components/headers/Sunset'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PluginListResponse'
              examples:
                catalogue:
                  $ref: '#/components/examples/PluginListExample'
        '500':
          $ref: '#/components/responses/InternalError'
      security: []

    post:
      tags:
        - Operator — Plugins
      summary: Publish first version of a new plugin
      description: |
        Creates a new plugin entry by publishing its first version (FR-OP-01). The plugin ID
        and version are extracted from the in-jar `marketplace-manifest.json`. The manifest
        `id` must match the intended plugin entry. Initial trust tier is set to `official`
        in Phase 1/2 (ADR-003).

        Publish bundle rules: required `.jar`; optional `CHANGELOG.md`; optional 0–5
        screenshot images (PNG/JPEG, max 2 MB each, NFR-08). See §6.3 in the marketplace plan.
      operationId: publishFirstVersion
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PublishBundleMultipart'
      responses:
        '201':
          description: Plugin published
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublishResponse'
              examples:
                published:
                  $ref: '#/components/examples/PublishResponseExample'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
      security:
        - operatorSessionJwt: []
        - githubOidcJwt: []
  /api/v1/plugins/{pluginId}:
    get:
      tags:
        - Public — Plugins
      summary: Get plugin detail
      description: |
        Returns plugin detail composed from operator state and the latest version manifest (§6.4).
        Returns `410 Gone` with a tombstone payload when the plugin has been removed (FR-OP-04).
      operationId: getPlugin
      parameters:
        - $ref: '#/components/parameters/PluginId'
      responses:
        '200':
          description: Plugin detail
          headers:
            Deprecation:
              $ref: '#/components/headers/Deprecation'
            Sunset:
              $ref: '#/components/headers/Sunset'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PluginDetail'
              examples:
                detail:
                  $ref: '#/components/examples/PluginDetailExample'
        '404':
          $ref: '#/components/responses/NotFound'
        '410':
          description: Plugin removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PluginTombstone'
              examples:
                removed:
                  $ref: '#/components/examples/PluginTombstoneExample'
        '500':
          $ref: '#/components/responses/InternalError'
      security: []

    patch:
      tags:
        - Operator — Plugins
      summary: Update operator-managed plugin fields
      description: |
        Updates operator-managed fields only (FR-OP-05). In-jar manifest fields cannot be
        edited on the registry side (ADR-002). In Phase 1/2 only `tier=official` is accepted;
        `partner` is schema-reserved until Phase 3 (ADR-003).
      operationId: updatePlugin
      parameters:
        - $ref: '#/components/parameters/PluginId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetTierRequest'
      responses:
        '200':
          description: Plugin updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PluginOperatorState'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
      security:
        - operatorSessionJwt: []
        - githubOidcJwt: []

    delete:
      tags:
        - Operator — Plugins
      summary: Remove a plugin entirely
      description: |
        Irreversible removal (FR-OP-04). Requires `removalReason` in the request body.
        Operator identity from the auth token is recorded as `removedBy`. Hard-deletes all
        version artifacts from storage and writes a tombstone for subsequent `410` responses.
      operationId: removePlugin
      parameters:
        - $ref: '#/components/parameters/PluginId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RemovePluginRequest'
      responses:
        '200':
          description: Plugin removed (tombstone confirmation)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PluginTombstone'
        '204':
          description: Plugin removed
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
      security:
        - operatorSessionJwt: []
        - githubOidcJwt: []
  /api/v1/plugins/{pluginId}/versions:
    get:
      tags:
        - Public — Plugins
      summary: List plugin versions
      description: |
        Returns version history with block state per version. Returns `410 Gone` when the
        plugin has been removed.
      operationId: listPluginVersions
      parameters:
        - $ref: '#/components/parameters/PluginId'
      responses:
        '200':
          description: Version history
          headers:
            Deprecation:
              $ref: '#/components/headers/Deprecation'
            Sunset:
              $ref: '#/components/headers/Sunset'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PluginVersionListResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '410':
          description: Plugin removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PluginTombstone'
        '500':
          $ref: '#/components/responses/InternalError'
      security: []

    post:
      tags:
        - Operator — Plugins
      summary: Publish a new version
      description: |
        Publishes a new immutable version for an existing plugin (FR-OP-02). The in-jar
        manifest `id` and `version` must match the URL path. Idempotent retry is supported
        when publish is interrupted before index commit (§6.4).
      operationId: publishVersion
      parameters:
        - $ref: '#/components/parameters/PluginId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PublishBundleMultipart'
      responses:
        '201':
          description: Version published
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublishResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
      security:
        - operatorSessionJwt: []
        - githubOidcJwt: []
  /api/v1/plugins/{pluginId}/versions/{version}:
    get:
      tags:
        - Public — Plugins
      summary: Get version detail
      description: |
        Returns composed version metadata: manifest fields, trust tier, block flag, optional
        security advisory (Phase 2), absolute CDN URLs for changelog and screenshots, and
        the SHA-256 checksum of the `.jar` artifact (used by `MarketplaceClient` before install).
      operationId: getPluginVersion
      parameters:
        - $ref: '#/components/parameters/PluginId'
        - $ref: '#/components/parameters/Version'
      responses:
        '200':
          description: Version detail
          headers:
            Deprecation:
              $ref: '#/components/headers/Deprecation'
            Sunset:
              $ref: '#/components/headers/Sunset'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PluginVersionDetail'
              examples:
                versionDetail:
                  $ref: '#/components/examples/PluginVersionDetailExample'
        '404':
          $ref: '#/components/responses/NotFound'
        '410':
          description: Plugin removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PluginTombstone'
        '500':
          $ref: '#/components/responses/InternalError'
      security: []
...
components:
  securitySchemes:
    operatorSessionJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Registry operator session token issued on login (ADR-010)
    githubOidcJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        GitHub Actions OIDC token for CI/CD publish (ADR-014). Distinguished from session
        JWTs by the `iss` claim (`token.actions.githubusercontent.com`).
    licenseJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Ed25519-signed license JWT for premium artifact download (ADR-012). Payload:
        `{ customerId, pluginId, iat, exp }` with 60 second expiry.
  parameters:
    PluginId:
      name: pluginId
      in: path
      required: true
      description: Plugin identifier from manifest (e.g. `plugin-jira-cloud`)
      schema:
        type: string
        pattern: '^[a-z0-9][a-z0-9-]{0,62}[a-z0-9]$'
        example: plugin-jira-cloud
    Version:
      name: version
      in: path
      required: true
      description: Semantic version (semver)
      schema:
        type: string
        pattern: '^(0|[1-9]\d*)\.(0|[1-9]\d*)\.(0|[1-9]\d*)(?:-[\w.-]+)?(?:\+[\w.-]+)?$'
        example: '1.4.2'
...
    CategoryQuery:
      name: category
      in: query
      required: false
      description: Filter by plugin category
      schema:
        $ref: '#/components/schemas/PluginCategory'
    SearchQuery:
      name: q
      in: query
      required: false
      description: Free-text search across plugin name and description
      schema:
        type: string
        minLength: 1
        maxLength: 200
...
  headers:
    Deprecation:
      description: |
        Present when API version `/api/v1/` is in deprecation window (NFR-15, ADR-006).
      schema:
        type: string
        example: 'true'
    Sunset:
      description: RFC 8594 sunset date for deprecated API version
      schema:
        type: string
        format: date-time
  responses:
    Unauthorized:
      description: Authentication required or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: UNAUTHORIZED
            message: Invalid or missing bearer token
    Forbidden:
      description: Authenticated but not permitted
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: FORBIDDEN
            message: Operator access denied
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: NOT_FOUND
            message: Plugin not found
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
          example:
            code: VALIDATION_ERROR
            message: Publish bundle validation failed
            errors:
              - field: manifest.id
                message: Manifest id does not match URL pluginId
    PayloadTooLarge:
      description: Publish bundle exceeds the maximum upload size accepted by the registry
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: PAYLOAD_TOO_LARGE
            message: Publish bundle exceeds the maximum upload size accepted by the registry
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: INTERNAL_ERROR
            message: Unexpected registry error
  schemas:
    PluginCategory:
      type: string
      enum:
        - bug-tracking
        - notifications
        - authorization
        - import
      description: RP-defined controlled vocabulary (§6.2)
    TrustTier:
      type: string
      enum:
        - official
        - partner
      description: |
        Operator-managed trust tier. Phase 1/2 accepts `official` only; `partner` is
        schema-reserved until Phase 3 (ADR-003).
    AccessTier:
      type: string
      enum:
        - public
        - premium
      default: public
      description: Author-declared access level from in-jar manifest (ADR-011)
    AdvisorySeverity:
      type: string
      enum:
        - low
        - medium
        - high
        - critical
    Author:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          example: ReportPortal Team
        email:
          type: string
          format: email
          example: support@reportportal.io
        url:
          type: string
          format: uri
          example: https://reportportal.io
    Compatibility:
      type: object
      required:
        - reportportal
      properties:
        reportportal:
          type: string
          description: Semver-style range against ReportPortal release version
          example: '>=25.1'
    PluginManifestFields:
      type: object
      required:
        - id
        - name
        - version
        - description
        - author
        - license
        - category
        - compatibility
      properties:
        id:
          type: string
          example: plugin-jira-cloud
        name:
          type: string
          example: Jira Cloud Integration
        version:
          type: string
          example: '1.4.2'
        description:
          type: string
          example: Post and link Jira Cloud issues from ReportPortal
        author:
          $ref: '#/components/schemas/Author'
        license:
          type: string
          example: Apache-2.0
        category:
          $ref: '#/components/schemas/PluginCategory'
        compatibility:
          $ref: '#/components/schemas/Compatibility'
        homepage:
          type: string
          format: uri
          example: https://reportportal.io/docs/plugins/jira-cloud
        access:
          $ref: '#/components/schemas/AccessTier'
        contactUrl:
          type: string
          format: uri
          description: Required when access is premium
          example: https://reportportal.io/pricing/service-packages/
    PluginListItem:
      type: object
      required:
        - id
        - name
        - latestVersion
        - category
        - access
        - tier
      properties:
        id:
          type: string
        name:
          type: string
        latestVersion:
          type: string
        description:
          type: string
        category:
          $ref: '#/components/schemas/PluginCategory'
        access:
          $ref: '#/components/schemas/AccessTier'
        tier:
          $ref: '#/components/schemas/TrustTier'
    PluginListResponse:
      type: object
      required:
        - plugins
      properties:
        plugins:
          type: array
          items:
            $ref: '#/components/schemas/PluginListItem'
    PluginDetail:
      allOf:
        - $ref: '#/components/schemas/PluginManifestFields'
        - type: object
          required:
            - tier
            - latestVersion
          properties:
            tier:
              $ref: '#/components/schemas/TrustTier'
            latestVersion:
              type: string
    PluginVersionSummary:
      type: object
      required:
        - version
        - blocked
      properties:
        version:
          type: string
        publishedAt:
          type: string
          format: date-time
          description: Timestamp when the version was first published
        blocked:
          type: boolean
        blockedAt:
          type: string
          format: date-time
          nullable: true
        blockReason:
          type: string
          nullable: true
    PluginVersionListResponse:
      type: object
      required:
        - pluginId
        - versions
      properties:
        pluginId:
          type: string
        versions:
          type: array
          items:
            $ref: '#/components/schemas/PluginVersionSummary'
    SecurityAdvisory:
      type: object
      required:
        - severity
        - text
        - attachedAt
      properties:
        severity:
          $ref: '#/components/schemas/AdvisorySeverity'
        text:
          type: string
          example: CVE-2026-1234 — vulnerable jackson-databind. Upgrade to 1.4.3.
        attachedAt:
          type: string
          format: date-time
    PluginVersionDetail:
      allOf:
        - $ref: '#/components/schemas/PluginManifestFields'
        - type: object
          required:
            - tier
            - blocked
            - sha256
            - screenshotUrls
          properties:
            tier:
              $ref: '#/components/schemas/TrustTier'
            blocked:
              type: boolean
            blockedAt:
              type: string
              format: date-time
              nullable: true
            blockReason:
              type: string
              nullable: true
            advisory:
              $ref: '#/components/schemas/SecurityAdvisory'
            sha256:
              type: string
              pattern: '^[a-f0-9]{64}$'
              description: SHA-256 checksum of the `.jar` artifact (hex lowercase)
              example: a1b2c3d4e5f6789012345678901234567890abcdef1234567890abcdef123456
            changelogUrl:
              type: string
              format: uri
              nullable: true
              description: Absolute CDN URL when CHANGELOG.md was uploaded; null otherwise
            screenshotUrls:
              type: array
              items:
                type: string
                format: uri
              description: Ordered absolute CDN URLs (alphabetical by filename)
    PluginTombstone:
      type: object
      required:
        - removed
        - removalReason
        - removedBy
      properties:
        removed:
          type: string
          format: date-time
        removalReason:
          type: string
        removedBy:
          type: string
          description: Operator identity from session JWT subject or OIDC subject
    BlockedVersion:
      type: object
      required:
        - version
        - blockedAt
        - reason
      properties:
        version:
          type: string
        blockedAt:
          type: string
          format: date-time
        reason:
          type: string
...
    PluginOperatorState:
      type: object
      required:
        - id
        - tier
        - latestVersion
      properties:
        id:
          type: string
        tier:
          $ref: '#/components/schemas/TrustTier'
        latestVersion:
          type: string
        blockedVersions:
          type: array
          items:
            $ref: '#/components/schemas/BlockedVersion'
    PublishResponse:
      type: object
      required:
        - pluginId
        - version
        - sha256
      properties:
        pluginId:
          type: string
        version:
          type: string
        sha256:
          type: string
          pattern: '^[a-f0-9]{64}$'
    PublishBundleMultipart:
      type: object
      required:
        - jar
      properties:
        jar:
          type: string
          format: binary
          description: Plugin `.jar` containing `marketplace-manifest.json`
        changelog:
          type: string
          format: binary
          description: Optional CHANGELOG.md (markdown)
        screenshots:
          type: array
          items:
            type: string
            format: binary
          maxItems: 5
          description: Optional PNG/JPEG screenshots (max 2 MB each, NFR-08)
...
    RemovePluginRequest:
      type: object
      required:
        - removalReason
      properties:
        removalReason:
          type: string
          minLength: 1
          maxLength: 2000
    SetTierRequest:
      type: object
      required:
        - tier
      properties:
        tier:
          $ref: '#/components/schemas/TrustTier'
...
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
    ValidationErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            errors:
              type: array
              items:
                type: object
                required:
                  - field
                  - message
                properties:
                  field:
                    type: string
                  message:
                    type: string
  examples:
    PluginListExample:
      value:
        plugins:
          - id: plugin-jira-cloud
            name: Jira Cloud Integration
            latestVersion: '1.4.2'
            description: Post and link Jira Cloud issues from ReportPortal
            category: bug-tracking
            access: public
            tier: official
          - id: plugin-quality-gate
            name: Quality Gate
            latestVersion: '2.0.0'
            description: Advanced quality gate rules
            category: import
            access: premium
            tier: official
    PluginDetailExample:
      value:
        id: plugin-jira-cloud
        name: Jira Cloud Integration
        version: '1.4.2'
        description: Post and link Jira Cloud issues from ReportPortal
        author:
          name: ReportPortal Team
          email: support@reportportal.io
          url: https://reportportal.io
        license: Apache-2.0
        category: bug-tracking
        compatibility:
          reportportal: '>=25.1'
        homepage: https://reportportal.io/docs/plugins/jira-cloud
        access: public
        tier: official
        latestVersion: '1.4.2'
    PluginVersionDetailExample:
      value:
        id: plugin-jira-cloud
        name: Jira Cloud Integration
        version: '1.4.2'
        description: Post and link Jira Cloud issues from ReportPortal
        author:
          name: ReportPortal Team
          email: support@reportportal.io
        license: Apache-2.0
        category: bug-tracking
        compatibility:
          reportportal: '>=25.1'
        homepage: https://reportportal.io/docs/plugins/jira-cloud
        access: public
        tier: official
        blocked: false
        sha256: a1b2c3d4e5f6789012345678901234567890abcdef1234567890abcdef123456
        changelogUrl: https://cdn.marketplace.reportportal.io/plugins/plugin-jira-cloud/versions/1.4.2/CHANGELOG.md
        screenshotUrls:
          - https://cdn.marketplace.reportportal.io/plugins/plugin-jira-cloud/versions/1.4.2/screenshots/01-main-view.png
    PluginTombstoneExample:
      value:
        removed: '2026-03-01T10:00:00Z'
        removalReason: Legal request — plugin withdrawn from catalogue
        removedBy: operator@reportportal.io
...
    PublishResponseExample:
      value:
        pluginId: plugin-jira-cloud
        version: '1.4.2'
        sha256: a1b2c3d4e5f6789012345678901234567890abcdef1234567890abcdef123456
...